requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
dateutils==0.6.12
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
idna==3.6
install==1.3.5
lxml==4.9.3
MouseInfo==0.1.3
multitasking==0.0.11
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.3.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
requests==2.31.0
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tesseract==0.1.3
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webencodings==0.5.1
wsproto==1.2.0
yfinance==0.2.28
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
cryptography==41.0.7
h11==0.14.0
idna==3.6
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pyOpenSSL==23.3.0
pytesseract==0.3.10
selenium==4.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
tesseract==0.1.3
trio==0.23.2
trio-websocket==0.11.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wsproto==1.2.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
dateutils==0.6.12
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
idna==3.6
install==1.3.5
lxml==4.9.3
MouseInfo==0.1.3
multitasking==0.0.11
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.3.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
requests==2.31.0
secure-smtplib==0.1.1
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tesseract==0.1.3
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webencodings==0.5.1
wsproto==1.2.0
yfinance==0.2.28
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
dateutils==0.6.12
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
idna==3.6
install==1.3.5
lxml==4.9.3
MouseInfo==0.1.3
multitasking==0.0.11
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.3.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
requests==2.31.0
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tesseract==0.1.3
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webencodings==0.5.1
wsproto==1.2.0
yfinance==0.2.28
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
dateutils==0.6.12
discord==2.3.2
discord.py==2.3.2
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozendict==2.3.8
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
idna==3.6
install==1.3.5
lxml==4.9.3
MouseInfo==0.1.3
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.3.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
responses==0.24.1
secure-smtplib==0.1.1
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tesseract==0.1.3
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.28
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
dateutils==0.6.12
discord==2.3.2
discord.py==2.3.2
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozendict==2.3.8
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
idna==3.6
install==1.3.5
lxml==4.9.3
MouseInfo==0.1.3
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.3.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
responses==0.24.1
secure-smtplib==0.1.1
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tesseract==0.1.3
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.28
